backend/requirements.txt
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
bcrypt==5.0.0
bidict==0.23.1
billiard==4.2.2
black==24.8.0
blinker==1.9.0
cachelib==0.13.0
celery==5.4.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.10.7
cryptography==46.0.2
dnspython==2.8.0
ecdsa==0.19.1
eventlet==0.36.1
fastapi==0.115.0
flake8==7.1.1
flasgger==0.9.7.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mistune==3.1.4
msgspec==0.19.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.3
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.14.1
PyYAML==6.0.3
redis==5.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.27.1
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.38.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.1
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
